server/requirements.txt
Flask
Flask-JWT-Extended
Flask-Bcrypt
python-dotenv
flask-cors



